requirements.txt
django==1.11.8
django_extensions
djangorestframework
django-storages
boto
psycopg2
uwsgi
django-push-notifications
